refactor: move LocalStream to dedicated module and remove unused headless routes
